+++ test/in-browser/spec/KefirSpecs.js
@@ -2215,6 +2215,12 @@ withTwoSources('bufferBy', extend({
 
   _handleSecondaryValue: function(x, isCurrent) {
     this._flush(isCurrent);
+  },
+
+  _handleSecondaryEnd: function(x, isCurrent) {
+    if (!this._flushOnEnd) {
+      this._send(END, null, isCurrent);
+    }
   }
 
 }, withTwoSourcesAndBufferMixin));
@@ -2222,16 +2228,22 @@ withTwoSources('bufferBy', extend({
 
 
 
-// withTwoSources('bufferWhileBy', extend({
+withTwoSources('bufferWhileBy', extend({
 
-//   _handlePrimaryValue: function(x, isCurrent) {
-//     this._buff.push(x);
-//     if (this._lastSecondary !== NOTHING && !this._lastSecondary) {
-//       this._flush(isCurrent);
-//     }
-//   }
+  _handlePrimaryValue: function(x, isCurrent) {
+    this._buff.push(x);
+    if (this._lastSecondary !== NOTHING && !this._lastSecondary) {
+      this._flush(isCurrent);
+    }
+  },
 
-// }, withTwoSourcesAndBufferMixin));
+  _handleSecondaryEnd: function(x, isCurrent) {
+    if (!this._flushOnEnd && (this._lastSecondary === NOTHING || this._lastSecondary)) {
+      this._send(END, null, isCurrent);
+    }
+  }
+
+}, withTwoSourcesAndBufferMixin));
 
 
 
@@ -21855,6 +21867,19 @@ describe('bufferBy', function() {
         return send(b, ['<end>']);
       });
     });
+    it('should do end when secondary ends if {flushOnEnd: false}', function() {
+      var a, b;
+      expect(stream().bufferBy(send(stream(), ['<end>']), {
+        flushOnEnd: false
+      })).toEmit(['<end:current>']);
+      a = stream();
+      b = stream();
+      return expect(a.bufferBy(b, {
+        flushOnEnd: false
+      })).toEmit(['<end>'], function() {
+        return send(b, ['<end>']);
+      });
+    });
     return it('should flush buffer (if not empty) on each value from secondary', function() {
       var a, b;
       a = stream();
@@ -21912,7 +21937,188 @@ describe('bufferBy', function() {
 
 
 
-},{"../test-helpers.coffee":91}],35:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],35:[function(require,module,exports){
+var Kefir, activate, deactivate, prop, send, stream, _ref;
+
+_ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir, deactivate = _ref.deactivate, activate = _ref.activate;
+
+describe('bufferWhileBy', function() {
+  describe('common', function() {
+    it('should activate/deactivate sources', function() {
+      var a, b;
+      a = stream();
+      b = stream();
+      expect(a.bufferWhileBy(b)).toActivate(a, b);
+      a = stream();
+      b = prop();
+      expect(a.bufferWhileBy(b)).toActivate(a, b);
+      a = prop();
+      b = stream();
+      expect(a.bufferWhileBy(b)).toActivate(a, b);
+      a = prop();
+      b = prop();
+      return expect(a.bufferWhileBy(b)).toActivate(a, b);
+    });
+    it('should end when primary ends', function() {
+      var a, b;
+      expect(send(stream(), ['<end>']).bufferWhileBy(stream())).toEmit(['<end:current>']);
+      a = stream();
+      b = stream();
+      return expect(a.bufferWhileBy(b)).toEmit(['<end>'], function() {
+        return send(a, ['<end>']);
+      });
+    });
+    it('should flush buffer on end', function() {
+      var a, b;
+      expect(send(prop(), [1, '<end>']).bufferWhileBy(stream())).toEmit([
+        {
+          current: [1]
+        }, '<end:current>'
+      ]);
+      a = stream();
+      b = stream();
+      return expect(a.bufferWhileBy(b)).toEmit([[1, 2], '<end>'], function() {
+        return send(a, [1, 2, '<end>']);
+      });
+    });
+    it('should not flush buffer on end if {flushOnEnd: false}', function() {
+      var a, b;
+      expect(send(prop(), [1, '<end>']).bufferWhileBy(stream(), {
+        flushOnEnd: false
+      })).toEmit(['<end:current>']);
+      a = stream();
+      b = stream();
+      return expect(a.bufferWhileBy(b, {
+        flushOnEnd: false
+      })).toEmit(['<end>'], function() {
+        return send(a, [1, 2, '<end>']);
+      });
+    });
+    it('should end when secondary ends, if it haven\'t emitted any value (w/ {flushOnEnd: false})', function() {
+      var a, b;
+      expect(stream().bufferWhileBy(send(stream(), ['<end>']), {
+        flushOnEnd: false
+      })).toEmit(['<end:current>']);
+      a = stream();
+      b = stream();
+      return expect(a.bufferWhileBy(b, {
+        flushOnEnd: false
+      })).toEmit(['<end>'], function() {
+        return send(b, ['<end>']);
+      });
+    });
+    it('should end when secondary ends, if its last emitted value was truthy (w/ {flushOnEnd: false})', function() {
+      var a, b;
+      expect(stream().bufferWhileBy(send(prop(), [true, '<end>']), {
+        flushOnEnd: false
+      })).toEmit(['<end:current>']);
+      a = stream();
+      b = stream();
+      return expect(a.bufferWhileBy(b, {
+        flushOnEnd: false
+      })).toEmit(['<end>'], function() {
+        return send(b, [true, '<end>']);
+      });
+    });
+    it('should not end when secondary ends, if its last emitted value was falsy (w/ {flushOnEnd: false})', function() {
+      var a, b;
+      expect(stream().bufferWhileBy(send(prop(), [false, '<end>']), {
+        flushOnEnd: false
+      })).toEmit([]);
+      a = stream();
+      b = stream();
+      return expect(a.bufferWhileBy(b, {
+        flushOnEnd: false
+      })).toEmit([], function() {
+        return send(b, [false, '<end>']);
+      });
+    });
+    it('should not end when secondary ends (w/o {flushOnEnd: false})', function() {
+      var a, b;
+      expect(stream().bufferWhileBy(send(prop(), ['<end>']))).toEmit([]);
+      a = stream();
+      b = stream();
+      expect(a.bufferWhileBy(b)).toEmit([], function() {
+        return send(b, ['<end>']);
+      });
+      expect(stream().bufferWhileBy(send(prop(), [true, '<end>']))).toEmit([]);
+      a = stream();
+      b = stream();
+      expect(a.bufferWhileBy(b)).toEmit([], function() {
+        return send(b, [true, '<end>']);
+      });
+      expect(stream().bufferWhileBy(send(prop(), [false, '<end>']))).toEmit([]);
+      a = stream();
+      b = stream();
+      return expect(a.bufferWhileBy(b)).toEmit([], function() {
+        return send(b, [false, '<end>']);
+      });
+    });
+    return it('should flush buffer on each value from primary if last value form secondary was falsy', function() {
+      var a, b;
+      a = stream();
+      b = stream();
+      return expect(a.bufferWhileBy(b)).toEmit([[1, 2, 3, 4], [5], [6, 7, 8]], function() {
+        send(a, [1, 2]);
+        send(b, [true]);
+        send(a, [3]);
+        send(b, [false]);
+        send(a, [4]);
+        send(a, [5]);
+        send(b, [true]);
+        send(a, [6, 7]);
+        send(b, [false]);
+        return send(a, [8]);
+      });
+    });
+  });
+  describe('stream + stream', function() {
+    return it('returns stream', function() {
+      return expect(stream().bufferWhileBy(stream())).toBeStream();
+    });
+  });
+  describe('stream + property', function() {
+    return it('returns stream', function() {
+      return expect(stream().bufferWhileBy(prop())).toBeStream();
+    });
+  });
+  describe('property + stream', function() {
+    it('returns property', function() {
+      return expect(prop().bufferWhileBy(stream())).toBeProperty();
+    });
+    return it('includes current to buffer', function() {
+      var a, b;
+      a = send(prop(), [1]);
+      b = stream();
+      return expect(a.bufferWhileBy(b)).toEmit([[1, 2]], function() {
+        send(b, [false]);
+        return send(a, [2]);
+      });
+    });
+  });
+  return describe('property + property', function() {
+    it('returns property', function() {
+      return expect(prop().bufferWhileBy(prop())).toBeProperty();
+    });
+    return it('both have current', function() {
+      var a, b;
+      a = send(prop(), [1]);
+      b = send(prop(), [false]);
+      expect(a.bufferWhileBy(b)).toEmit([
+        {
+          current: [1]
+        }
+      ]);
+      a = send(prop(), [1]);
+      b = send(prop(), [true]);
+      return expect(a.bufferWhileBy(b)).toEmit([]);
+    });
+  });
+});
+
+
+
+},{"../test-helpers.coffee":92}],36:[function(require,module,exports){
 var Kefir, not3, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -22003,7 +22209,7 @@ describe('bufferWhile', function() {
 
 
 
-},{"../test-helpers.coffee":91}],36:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],37:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -22139,7 +22345,7 @@ describe('bus', function() {
 
 
 
-},{"../test-helpers.coffee":91}],37:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],38:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -22174,7 +22380,7 @@ describe('changes', function() {
 
 
 
-},{"../test-helpers.coffee":91}],38:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],39:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref,
   __slice = [].slice;
 
@@ -22288,7 +22494,7 @@ describe('combine', function() {
 
 
 
-},{"../test-helpers.coffee":91}],39:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],40:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -22396,7 +22602,7 @@ describe('concat', function() {
 
 
 
-},{"../test-helpers.coffee":91}],40:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],41:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -22416,7 +22622,7 @@ describe('constant', function() {
 
 
 
-},{"kefir":93}],41:[function(require,module,exports){
+},{"kefir":94}],42:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -22607,7 +22813,7 @@ describe('debounce', function() {
 
 
 
-},{"../test-helpers.coffee":91}],42:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],43:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -22671,7 +22877,7 @@ describe('delay', function() {
 
 
 
-},{"../test-helpers.coffee":91}],43:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],44:[function(require,module,exports){
 var Kefir, minus, noop, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -22771,7 +22977,7 @@ describe('diff', function() {
 
 
 
-},{"../test-helpers.coffee":91}],44:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],45:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -22797,7 +23003,7 @@ describe('emitter', function() {
 
 
 
-},{"kefir":93}],45:[function(require,module,exports){
+},{"kefir":94}],46:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -23046,7 +23252,7 @@ describe('filterBy', function() {
 
 
 
-},{"../test-helpers.coffee":91}],46:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],47:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -23133,7 +23339,7 @@ describe('filter', function() {
 
 
 
-},{"../test-helpers.coffee":91}],47:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],48:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -23268,7 +23474,7 @@ describe('flatMapConcat', function() {
 
 
 
-},{"../test-helpers.coffee":91}],48:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],49:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -23423,7 +23629,7 @@ describe('flatMapFirst', function() {
 
 
 
-},{"../test-helpers.coffee":91}],49:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],50:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -23557,7 +23763,7 @@ describe('flatMapLatest', function() {
 
 
 
-},{"../test-helpers.coffee":91}],50:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],51:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -23688,7 +23894,7 @@ describe('flatMapConcurLimit', function() {
 
 
 
-},{"../test-helpers.coffee":91}],51:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],52:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -23851,7 +24057,7 @@ describe('flatMap', function() {
 
 
 
-},{"../test-helpers.coffee":91}],52:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],53:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -23969,7 +24175,7 @@ describe('flatten', function() {
 
 
 
-},{"../test-helpers.coffee":91}],53:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],54:[function(require,module,exports){
 var Kefir, activate, deactivate, _ref;
 
 _ref = require('../test-helpers.coffee'), activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -24051,7 +24257,7 @@ describe('fromBinder', function() {
 
 
 
-},{"../test-helpers.coffee":91}],54:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],55:[function(require,module,exports){
 var Kefir, activate, deactivate, _ref;
 
 _ref = require('../test-helpers.coffee'), activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -24114,7 +24320,7 @@ describe('fromCallback', function() {
 
 
 
-},{"../test-helpers.coffee":91}],55:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],56:[function(require,module,exports){
 var Kefir, activate, deactivate, _ref;
 
 _ref = require('../test-helpers.coffee'), activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -24247,7 +24453,7 @@ describe('fromEvent', function() {
 
 
 
-},{"../test-helpers.coffee":91}],56:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],57:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -24267,7 +24473,7 @@ describe('fromPoll', function() {
 
 
 
-},{"kefir":93}],57:[function(require,module,exports){
+},{"kefir":94}],58:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -24283,7 +24489,7 @@ describe('interval', function() {
 
 
 
-},{"kefir":93}],58:[function(require,module,exports){
+},{"kefir":94}],59:[function(require,module,exports){
 var $, Kefir, countListentrs, inBrowser, withDOM, _ref;
 
 _ref = require('../test-helpers.coffee'), withDOM = _ref.withDOM, inBrowser = _ref.inBrowser, Kefir = _ref.Kefir;
@@ -24593,7 +24799,7 @@ if (!inBrowser) {
 
 
 
-},{"../test-helpers.coffee":91,"addons/kefir-jquery":92,"jquery":6}],59:[function(require,module,exports){
+},{"../test-helpers.coffee":92,"addons/kefir-jquery":93,"jquery":6}],60:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -24609,7 +24815,7 @@ describe('later', function() {
 
 
 
-},{"kefir":93}],60:[function(require,module,exports){
+},{"kefir":94}],61:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -24686,7 +24892,7 @@ describe('mapEnd', function() {
 
 
 
-},{"../test-helpers.coffee":91}],61:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],62:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -24744,7 +24950,7 @@ describe('map', function() {
 
 
 
-},{"../test-helpers.coffee":91}],62:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],63:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -24840,7 +25046,7 @@ describe('merge', function() {
 
 
 
-},{"../test-helpers.coffee":91}],63:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],64:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -24856,7 +25062,7 @@ describe('never', function() {
 
 
 
-},{"kefir":93}],64:[function(require,module,exports){
+},{"kefir":94}],65:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, activate = _ref.activate, deactivate = _ref.deactivate, Kefir = _ref.Kefir;
@@ -24955,7 +25161,7 @@ describe('pool', function() {
 
 
 
-},{"../test-helpers.coffee":91}],65:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],66:[function(require,module,exports){
 var Kefir, activate, prop, send, _ref;
 
 _ref = require('../test-helpers.coffee'), prop = _ref.prop, send = _ref.send, activate = _ref.activate, Kefir = _ref.Kefir;
@@ -25115,7 +25321,7 @@ describe('Property', function() {
 
 
 
-},{"../test-helpers.coffee":91}],66:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],67:[function(require,module,exports){
 var Kefir, minus, noop, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -25211,7 +25417,7 @@ describe('reduce', function() {
 
 
 
-},{"../test-helpers.coffee":91}],67:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],68:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -25230,7 +25436,7 @@ describe('repeatedly', function() {
 
 
 
-},{"kefir":93}],68:[function(require,module,exports){
+},{"kefir":94}],69:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref,
   __slice = [].slice;
 
@@ -25367,7 +25573,7 @@ describe('sampledBy', function() {
 
 
 
-},{"../test-helpers.coffee":91}],69:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],70:[function(require,module,exports){
 var Kefir, minus, noop, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -25457,7 +25663,7 @@ describe('scan', function() {
 
 
 
-},{"../test-helpers.coffee":91}],70:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],71:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -25476,7 +25682,7 @@ describe('sequentially', function() {
 
 
 
-},{"kefir":93}],71:[function(require,module,exports){
+},{"kefir":94}],72:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -25552,7 +25758,7 @@ describe('skipDuplicates', function() {
 
 
 
-},{"../test-helpers.coffee":91}],72:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],73:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -25611,7 +25817,7 @@ describe('mapEnd', function() {
 
 
 
-},{"../test-helpers.coffee":91}],73:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],74:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir, activate = _ref.activate, deactivate = _ref.deactivate;
@@ -25865,7 +26071,7 @@ describe('skipUntilBy', function() {
 
 
 
-},{"../test-helpers.coffee":91}],74:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],75:[function(require,module,exports){
 var Kefir, activate, deactivate, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir, deactivate = _ref.deactivate, activate = _ref.activate;
@@ -26122,7 +26328,7 @@ describe('skipWhileBy', function() {
 
 
 
-},{"../test-helpers.coffee":91}],75:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],76:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -26250,7 +26456,7 @@ describe('skipWhile', function() {
 
 
 
-},{"../test-helpers.coffee":91}],76:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],77:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -26350,7 +26556,7 @@ describe('skip', function() {
 
 
 
-},{"../test-helpers.coffee":91}],77:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],78:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -26446,7 +26652,7 @@ describe('slidingWindow', function() {
 
 
 
-},{"../test-helpers.coffee":91}],78:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],79:[function(require,module,exports){
 var Kefir, activate, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, send = _ref.send, activate = _ref.activate, Kefir = _ref.Kefir;
@@ -26637,7 +26843,7 @@ describe('Stream', function() {
 
 
 
-},{"../test-helpers.coffee":91}],79:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],80:[function(require,module,exports){
 var Kefir, expectToBehaveAsMap, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -26889,7 +27095,7 @@ describe('awaiting', function() {
 
 
 
-},{"../test-helpers.coffee":91}],80:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],81:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -27133,7 +27339,7 @@ describe('takeUntilBy', function() {
 
 
 
-},{"../test-helpers.coffee":91}],81:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],82:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -27382,7 +27588,7 @@ describe('takeWhileBy', function() {
 
 
 
-},{"../test-helpers.coffee":91}],82:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],83:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -27519,7 +27725,7 @@ describe('takeWhile', function() {
 
 
 
-},{"../test-helpers.coffee":91}],83:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],84:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -27604,7 +27810,7 @@ describe('take', function() {
 
 
 
-},{"../test-helpers.coffee":91}],84:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],85:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -27878,7 +28084,7 @@ describe('throttle', function() {
 
 
 
-},{"../test-helpers.coffee":91}],85:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],86:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -27967,7 +28173,7 @@ describe('timestamp', function() {
 
 
 
-},{"../test-helpers.coffee":91}],86:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],87:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -28013,7 +28219,7 @@ describe('toProperty', function() {
 
 
 
-},{"../test-helpers.coffee":91}],87:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],88:[function(require,module,exports){
 var Kefir, comp, noop, prop, send, stream, testWithLib, _ref,
   __slice = [].slice;
 
@@ -28263,7 +28469,7 @@ describe('transduce', function() {
 
 
 
-},{"../test-helpers.coffee":91,"transducers-js":32,"transducers.js":33}],88:[function(require,module,exports){
+},{"../test-helpers.coffee":92,"transducers-js":32,"transducers.js":33}],89:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -28342,7 +28548,7 @@ describe('withDefault', function() {
 
 
 
-},{"../test-helpers.coffee":91}],89:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],90:[function(require,module,exports){
 var Kefir, prop, send, stream, _ref;
 
 _ref = require('../test-helpers.coffee'), stream = _ref.stream, prop = _ref.prop, send = _ref.send, Kefir = _ref.Kefir;
@@ -28459,7 +28665,7 @@ describe('withHandler', function() {
 
 
 
-},{"../test-helpers.coffee":91}],90:[function(require,module,exports){
+},{"../test-helpers.coffee":92}],91:[function(require,module,exports){
 var Kefir;
 
 Kefir = require('kefir');
@@ -28485,7 +28691,7 @@ describe('withInterval', function() {
 
 
 
-},{"kefir":93}],91:[function(require,module,exports){
+},{"kefir":94}],92:[function(require,module,exports){
 var Kefir, getCurrent, logItem, sinon, _activateHelper,
   __slice = [].slice;
 
@@ -28734,7 +28940,7 @@ beforeEach(function() {
 
 
 
-},{"../dist/kefir":1,"sinon":7}],92:[function(require,module,exports){
+},{"../dist/kefir":1,"sinon":7}],93:[function(require,module,exports){
 /*! An addon for Kefir.js v0.4.0
  *  https://github.com/pozadi/kefir
  */
@@ -28786,6 +28992,6 @@ beforeEach(function() {
 
 }(this));
 
-},{"jquery":6,"kefir":93}],93:[function(require,module,exports){
+},{"jquery":6,"kefir":94}],94:[function(require,module,exports){
 module.exports=require(1)
-},{"/Users/anon/projects/my/kefir/dist/kefir.js":1}]},{},[34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90]);
+},{"/Users/anon/projects/my/kefir/dist/kefir.js":1}]},{},[34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91]);